refactor: 추후 삭제할 DetailModal, ConfirmModal을 apps 폴더로 이동
AlertModal은 참조되는 곳이 없어 삭제
