+++ apps/client/src/pages/new-memo/components/memo-input/memo-input.tsx
@@ -1,9 +1,10 @@
 import { ChangeEvent, useMemo, useState } from 'react';
 
-import { Button, ConfirmModal } from '@cds/ui';
+import { Button } from '@cds/ui';
 
 import { htmlToMarkdown } from '@pages/new-memo/utils/html-to-markdown';
 
+import ConfirmModal from '@shared/components/modals/confirm-modal/confirm-modal';
 import { LabelTextType } from '@shared/types/label-type';
 
 import { useCreateMemo } from '../../apis/queries';
