+++ apps/client/src/shared/components/modals/detail-modal/detail-modal.css.ts
@@ -1,7 +1,7 @@
 import { style } from '@vanilla-extract/css';
 import { recipe } from '@vanilla-extract/recipes';
 
-import { themeVars } from '../../styles';
+import { themeVars } from '../../../../../../../packages/cds-ui/src/styles';
 
 export const container = recipe({
   base: {
